don't parse url for getServiceURL unless required
fixes https://github.com/cloudfoundry-community/node-cfenv/issues/21

Turns out the following is not always true:

```js
  url.format(url.parse(SomeURL)) == SomeURL
```

At least when the protocol isn't `http:` or `https:`.  cfenv expects this
to be true in `getServiceURL()`, when using the replacements argument.  It
seems to break for some example mongodb urls, like the one added to the
test in this commit.

Fix for now is that if there aren't any replacement values, then return early
with the currently calculated URL.  So it would still break if you passed
one of these not-parsed-correctly URLs and replacements, but the odds of that
seem pretty slim.

Also noticed that `url.format()`` must be standardizing URLs to return them with
a trailing `/` if there was no other path.  So now you don't always get that
trailing `/`. Seems survivable.