handle ports race condition by returning 3000
fixes https://github.com/cloudfoundry-community/node-cfenv/issues/40

This is a very simple to fix to the problem with ports where it
will occaisonally throw an error:

- https://github.com/hoodiehq-archive/node-ports/issues/4

The fix is to catch the error and use port 3000.  Not great, but should
handle some cases where this was a problem.

An alternate fix to try would be to try fetching the port again, probably
in a loop (with some limit), in hopes of getting a value from ports.